add confirmation when deleting book from list | add correcy queryselector params to get span next to header | some refactoring

+++ public/main.js
@@ -5,8 +5,7 @@
     var comm         = new Comm({ endpoint: "" }),
         doc          = window.document,
         queryButton  = doc.getElementById('addBookBtn'),
-        openIsbnForm = doc.querySelector('.icon-barcode'),
-        openCustForm = doc.querySelector('.icon-pencil'),
+        openIsbnForm = doc.querySelector('.openform'),
         form         = doc.getElementById("addBook"),
         table        = doc.querySelector('table'),
         previewBox   = doc.querySelector('.preview-box'),
@@ -39,19 +38,22 @@
     function deleteBook(e) {
         e.stopPropagation();
         e.preventDefault();
-        var bookID = e.target.rel;
-        comm.deleteBook(bookID,
-            function(rsp) {
-                updateBooklist(rsp, "delete");
-            },
-            function(err) {
-            }
-        );
+        var confirmation = confirm('Are you sure you want to delete this book?');
+        if(confirmation) {
+            var bookID = e.target.parentElement.rel;
+            comm.deleteBook(bookID,
+                function(rsp) {
+                    updateBooklist(rsp, "delete");
+                },
+                function(err) {
+                }
+            );
+        }
     }
 
     function updateBooklist(data, option){
         var bookTable = doc.querySelector('.list-container tbody');
-        var bookNumber = doc.querySelector('header span');
+        var bookNumber = doc.querySelector('header h1 span');
         bookTable.innerHTML = data;
         if("add" === option) {
             bookNumber.innerHTML = parseInt(bookNumber.innerHTML, 10)+1;
@@ -107,18 +109,18 @@
     }
 
     function bookPreview(previewJSON) {
-        var previewInfo  = JSON.parse(previewJSON),
-            title = previewInfo.title,
-            author = previewInfo.author,
-            thumb = previewInfo.coverLink.thumbnail,
-            date = previewInfo.date,
-            desc = previewInfo.description,
-            imgContainer  = previewBox.querySelector('.img-container'),
-            descContainer = previewBox.querySelector('.description'),
-            myPreviewImg  = imgContainer.cloneNode(true),
-            myPreviewDesc = descContainer.cloneNode(true);
+        var previewInfo     = JSON.parse(previewJSON),
+            title           = previewInfo.title,
+            author          = previewInfo.author,
+            thumb           = previewInfo.coverLink.thumbnail,
+            date            = previewInfo.date,
+            desc            = previewInfo.description,
+            imgContainer    = previewBox.querySelector('.img-container'),
+            descContainer   = previewBox.querySelector('.description'),
+            myPreviewImg    = imgContainer.cloneNode(true),
+            myPreviewDesc   = descContainer.cloneNode(true);
 
-            myPreviewImg.innerHTML = '<img class="img-polaroid" src='+thumb+'>';
+            myPreviewImg.innerHTML = '<img class="img-thumbnail" src='+thumb+'>';
             myPreviewDesc.innerHTML = '<h4>'+title+'</h4><p>'+author+'</p><p>'+date+'</p><p>'+desc.slice(0,600)+'...</p>';
 
         imgContainer.parentNode.replaceChild(myPreviewImg, imgContainer);
@@ -133,20 +135,28 @@
     }
 
 
-    var revealForm = function(e) {
-        form.classList.toggle('hide');
+    var revealForm = (function(e) {
+        // little fucked up closure for button inner text magic...
+        var buttonPhrase = "Nevermind... Let me see the list";
+        return function(e) {
+            var interimVar = e.target.innerHTML;
+            e.target.innerHTML = buttonPhrase;
+            buttonPhrase = interimVar;
 
-        if (e.target.className === "icon-barcode"){
-            form.querySelector("#author").parentNode.classList.add('hide');
-            form.querySelector("#title").parentNode.classList.add('hide');
-            form.querySelector("#isbn").parentNode.classList.remove('hide');
-            dater();
-        } else {
-            form.querySelector("#isbn").parentNode.classList.add('hide');
-            form.querySelector("#author").parentNode.classList.remove('hide');
-            form.querySelector("#title").parentNode.classList.remove('hide');
+            form.classList.toggle('hide');
+
+            if (e.target.className !== "icon-barcode"){
+                form.querySelector("#author").parentNode.classList.add('hide');
+                form.querySelector("#title").parentNode.classList.add('hide');
+                form.querySelector("#isbn").parentNode.classList.remove('hide');
+                dater();
+            } else {
+                form.querySelector("#isbn").parentNode.classList.add('hide');
+                form.querySelector("#author").parentNode.classList.remove('hide');
+                form.querySelector("#title").parentNode.classList.remove('hide');
+            }
         }
-    };
+    }());
 
 
 
@@ -157,7 +167,6 @@
 
     queryButton.addEventListener('click', queryBook, false);
     openIsbnForm.addEventListener('click', revealForm, false);
-    //openCustForm.addEventListener('click', revealForm, false);
     bookAccept.addEventListener('click', addBook, false);
     bookDecline.addEventListener('click', declineBook, false);
     table.addEventListener('click', deleteBook, false);
